Revert "Bratseth/iterate over indexes not fields 2"

+++ container-search/src/main/java/com/yahoo/prelude/searcher/FieldCollapsingSearcher.java
@@ -19,9 +19,10 @@ import java.util.Map;
 
 
 /**
- * A searcher which does parametrized collapsing.
+ * A searcher which does parametrized collapsing. Based on
+ * SiteCollapsingSearcher. Deprecated - use grouping.
  *
- * @author Steinar Knutsen
+ * @author  <a href="mailto:steinar@yahoo-inc.com">Steinar Knutsen</a>
  */
 @SuppressWarnings("deprecation")
 @After(PhaseNames.RAW_QUERY)
@@ -173,18 +174,17 @@ public class FieldCollapsingSearcher extends Searcher {
             }
 
             if (knownCollapses.containsKey(collapseId)) {
-                int numHitsThisField = knownCollapses.get(collapseId);
+                int numHitsThisField = knownCollapses.get(collapseId).intValue();
 
                 if (numHitsThisField < collapseSize) {
                     result.hits().add(hit);
                     ++numHitsThisField;
-                    knownCollapses.put(collapseId, numHitsThisField);
+                    knownCollapses.put(collapseId, Integer.valueOf(numHitsThisField));
                 }
             } else {
-                knownCollapses.put(collapseId, 1);
+                knownCollapses.put(collapseId, Integer.valueOf(1));
                 result.hits().add(hit);
             }
         }
     }
-
 }
